modified:   Customer.py 	modified:   DB.json 	modified:   DesignClass.py 	modified:   __pycache__/Customer.cpython-310.pyc 	modified:   __pycache__/DesignClass.cpython-310.pyc 	modified:   __pycache__/intro_bank.cpython-310.pyc 	modified:   __pycache__/loan_window.cpython-310.pyc 	modified:   __pycache__/login.cpython-310.pyc 	modified:   __pycache__/ui_main.cpython-310.pyc 	modified:   bank.py 	new file:   customer_graphs/ETR-666634-104277.png 	new file:   customer_graphs/ETR-885933-558214.png 	new file:   customer_reports/asd_4444444444444.pdf 	modified:   icons/logo_title.png 	modified:   lastwilltolive.ipynb 	modified:   loan_window.py 	modified:   login.py 	modified:   ui_main.py

+++ Customer.py
@@ -23,7 +23,7 @@ class Customer:
         current_date = date.today()
         formatted_date = current_date.strftime("%d/%m/%Y")
         try:
-            amount = int(self.deposit_edit.text())
+            amount = float(self.deposit_edit.text())
             self.deposit_edit.setText("")
         except:
             print("NAN")
